Move xpath func test to correct file
